Remove decimal place from status bar output
Centre point is always an integer. Here we are reporting the starting
centre point (the seed point) for each ellipsoid, not the slightly
offset centroid of the fitted maximal ellipsoid, which would be returned
by QuickEllipsoid.getCentre().

+++ Modern/ops/src/main/java/org/bonej/ops/ellipsoid/EllipsoidOptimisationStrategy.java
@@ -543,7 +543,7 @@ public class EllipsoidOptimisationStrategy extends AbstractBinaryFunctionOp<byte
 		logService.debug("Optimised ellipsoid in " + (stop - start) + " ms after " + totalIterations + " iterations ("
 				+ (double) (stop - start) / totalIterations + " ms/iteration)");
 
-		String centreString = "("+centre[0]+",  "+centre[1]+",  "+centre[2]+")";
+		String centreString = "("+(int) centre[0]+",  "+ (int) centre[1]+",  "+(int) centre[2]+")";
 		if(statusService!=null) {
 			statusService.showStatus("Ellipsoid optimised at " + centreString);//non-null check needed for tests
 		}
